// web/src/gen/pb/api/v1/data.ts
export interface ListDashboardSummariesResponse {
    /**
     * List of dashboard summaries.
     *
     * @generated from protobuf field: repeated testgrid.api.v1.DashboardSummary dashboard_summaries = 1
     */
    dashboardSummaries: DashboardSummary[];
}
// ...
export interface DashboardSummary {
    /**
     * Name of the dashboard.
     *
     * @generated from protobuf field: string name = 1
     */
    name: string;
    /**
     * Overall status of the dashboard.
     * Will be calculated based on the presence and importance of underlying tabs.
     *
     * @generated from protobuf field: string overall_status = 2
     */
    overallStatus: string;
    /**
     * Count of the tabs by status.
     *
     * @generated from protobuf field: map<string, int32> tab_status_count = 3
     */
    tabStatusCount: {
        [key: string]: number;
    };
}
// ...
class ListDashboardSummariesResponse$Type extends MessageType<ListDashboardSummariesResponse> {
    constructor() {
        super("testgrid.api.v1.ListDashboardSummariesResponse", [
            { no: 1, name: "dashboard_summaries", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => DashboardSummary }
        ]);
    }
    create(value?: PartialMessage<ListDashboardSummariesResponse>): ListDashboardSummariesResponse {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.dashboardSummaries = [];
        if (value !== undefined)
            reflectionMergePartial<ListDashboardSummariesResponse>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ListDashboardSummariesResponse): ListDashboardSummariesResponse {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* repeated testgrid.api.v1.DashboardSummary dashboard_summaries */ 1:
                    message.dashboardSummaries.push(DashboardSummary.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ListDashboardSummariesResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* repeated testgrid.api.v1.DashboardSummary dashboard_summaries = 1; */
        for (let i = 0; i < message.dashboardSummaries.length; i++)
            DashboardSummary.internalBinaryWrite(message.dashboardSummaries[i], writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message testgrid.api.v1.ListDashboardSummariesResponse
 */
export const ListDashboardSummariesResponse = new ListDashboardSummariesResponse$Type();
// ...
class DashboardSummary$Type extends MessageType<DashboardSummary> {
    constructor() {
        super("testgrid.api.v1.DashboardSummary", [
            { no: 1, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "overall_status", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "tab_status_count", kind: "map", K: 9 /*ScalarType.STRING*/, V: { kind: "scalar", T: 5 /*ScalarType.INT32*/ } }
        ]);
    }
    create(value?: PartialMessage<DashboardSummary>): DashboardSummary {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.name = "";
        message.overallStatus = "";
        message.tabStatusCount = {};
        if (value !== undefined)
            reflectionMergePartial<DashboardSummary>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: DashboardSummary): DashboardSummary {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string name */ 1:
                    message.name = reader.string();
                    break;
                case /* string overall_status */ 2:
                    message.overallStatus = reader.string();
                    break;
                case /* map<string, int32> tab_status_count */ 3:
                    this.binaryReadMap3(message.tabStatusCount, reader, options);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    private binaryReadMap3(map: DashboardSummary["tabStatusCount"], reader: IBinaryReader, options: BinaryReadOptions): void {
        let len = reader.uint32(), end = reader.pos + len, key: keyof DashboardSummary["tabStatusCount"] | undefined, val: DashboardSummary["tabStatusCount"][any] | undefined;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case 1:
                    key = reader.string();
                    break;
                case 2:
                    val = reader.int32();
                    break;
                default: throw new globalThis.Error("unknown map entry field for testgrid.api.v1.DashboardSummary.tab_status_count");
            }
        }
        map[key ?? ""] = val ?? 0;
    }
    internalBinaryWrite(message: DashboardSummary, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string name = 1; */
        if (message.name !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.name);
        /* string overall_status = 2; */
        if (message.overallStatus !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.overallStatus);
        /* map<string, int32> tab_status_count = 3; */
        for (let k of globalThis.Object.keys(message.tabStatusCount))
            writer.tag(3, WireType.LengthDelimited).fork().tag(1, WireType.LengthDelimited).string(k).tag(2, WireType.Varint).int32(message.tabStatusCount[k]).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message testgrid.api.v1.DashboardSummary
 */
export const DashboardSummary = new DashboardSummary$Type();